backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-chess>=1.999
requests>=2.31.0
